requirements.txt
# ============================================================
# LLM Optimization Pipeline — Requirements
# Install: pip install -r requirements.txt
# ============================================================

# --- Step 1: Download model from HuggingFace ---
huggingface-hub

# --- Step 2: Generate training dataset (8B teacher) ---
transformers==4.57.6
datasets
requests                  # call Ollama API to generate examples
pyyaml                    # read config.yaml
tqdm                      # progress bars with ETA

# --- Step 3A: Structured Pruning (Path A) ---
torch
torch-pruning

# --- Step 3B: Knowledge Distillation (Path B) ---
unsloth==2026.4.4         # 2x faster fine-tuning, 60% less VRAM

# --- Step 4A/4B: Fine-tuning LoRA/QLoRA ---
peft==0.18.1              # LoRA / QLoRA adapters
trl==0.24.0              # SFTTrainer for supervised fine-tuning
bitsandbytes==0.49.2      # 4-bit quantization for QLoRA (CUDA only)
accelerate==1.13.0        # distributed training / mixed precision

# --- Step 5+6: OpenVINO conversion + NNCF quantization ---
openvino
optimum-intel             # HuggingFace ↔ OpenVINO bridge
nncf                      # Neural Network Compression (INT4 AWQ)

# --- Step 10: Benchmark (Flywheel + Guardrails) ---
matplotlib                # Pareto plot
numpy
jsonschema                # JSON schema validation for guardrails
